export our base redux tooling to a core package
