+++ packages/core/constants.js
@@ -1,0 +1,2 @@
+// @flow
+module.exports = require("./").constants;
