+++ packages/core/package.json
@@ -13,6 +13,8 @@
     "build:lib:watch": "npm run build:lib -- --watch"
   },
   "dependencies": {
+    "@nteract/commutable": "^2.1.2",
+    "escape-carriage": "^1.2.0",
     "rxjs": "^5.5.0",
     "uuid": "^3.1.0"
   },
