+++ packages/core/records.js
@@ -1,0 +1,2 @@
+// @flow
+module.exports = require("./").records;
