+++ packages/core/src/index.js
@@ -1,3 +1,9 @@
 //@flow
 
-// Nothing here yet
+import * as actions from "./actions";
+import * as constants from "./constants";
+import * as middlewares from "./middlewares";
+import * as records from "./records";
+import * as reducers from "./reducers";
+
+export { actions, constants, middlewares, records, reducers };
