+++ packages/core/src/reducers/config.js
