+++ packages/core/src/reducers/document.js
