+++ packages/core/src/reducers/index.js
@@ -1,0 +1,8 @@
+// @flow
+import document from "./document";
+import metadata from "./metadata";
+import comms from "./comms";
+import config from "./config";
+
+// We'll export our specific reducers here
+export { metadata, document, comms, config };
