+++ packages/core/src/reducers/metadata.js
