+++ packages/desktop/src/notebook/components/notebook.js
@@ -16,7 +16,7 @@ import {
   focusNextCellEditor,
   moveCell,
   executeCell
-} from "../actions";
+} from "@nteract/core/actions";
 import type { CellProps } from "./cell/cell";
 
 // NOTE: PropTypes are required for the sake of contextTypes
