+++ packages/desktop/src/notebook/middlewares.js
@@ -1,4 +1,6 @@
 // @flow
+import { errorMiddleware } from "@nteract/core/middlewares";
+
 import { createEpicMiddleware, combineEpics } from "redux-observable";
 
 import epics from "./epics";
@@ -9,33 +11,12 @@ type Action = {
   type: string
 };
 
-export const errorMiddleware = (store: any) => (next: any) => (
-  action: Action
-) => {
-  if (!action.type.includes("ERROR")) {
-    return next(action);
-  }
-  console.error(action);
-  let errorText;
-  if (action.payload) {
-    errorText = JSON.stringify(action.payload, null, 2);
-  } else {
-    errorText = JSON.stringify(action, null, 2);
-  }
-  const state = store.getState();
-  const notificationSystem = state.app.get("notificationSystem");
-  if (notificationSystem) {
-    notificationSystem.addNotification({
-      title: action.type,
-      message: errorText,
-      dismissible: true,
-      position: "tr",
-      level: "error"
-    });
-  }
-  return next(action);
-};
-
 const middlewares = [createEpicMiddleware(rootEpic), errorMiddleware];
+
+if (process.env.DEBUG === "true") {
+  const logger = require("./logger"); // eslint-disable-line global-require
+
+  middlewares.push(logger());
+}
 
 export default middlewares;
