+++ packages/desktop/src/notebook/providers/editor.js
@@ -2,7 +2,12 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
 
-import { focusCell, focusCellEditor, updateCellSource } from "../actions";
+import {
+  focusCell,
+  focusCellEditor,
+  updateCellSource
+} from "@nteract/core/actions";
+
 import EditorView from "@nteract/editor/";
 
 type Props = {
