+++ packages/desktop/src/notebook/reducers/index.js
@@ -1,11 +1,8 @@
 // @flow
 import { combineReducers } from "redux";
 
+import { document, metadata, comms, config } from "@nteract/core/reducers";
 import app from "./app";
-import document from "./document";
-import metadata from "./metadata";
-import comms from "./comms";
-import config from "./config";
 
 export default combineReducers({
   app,
