+++ packages/desktop/src/notebook/store.js
@@ -1,18 +1,13 @@
 /* @flow */
-import { createStore, applyMiddleware } from "redux";
+import { createStore, applyMiddleware, combineReducers } from "redux";
 
-import type { AppState } from "./records";
+import type { AppState } from "@nteract/core/records";
 
 import middlewares from "./middlewares";
 import rootReducer from "./reducers";
 
-/* istanbul ignore if -- only used for debugging */
-if (process.env.DEBUG === "true") {
-  const logger = require("./logger"); // eslint-disable-line global-require
-
-  middlewares.push(logger());
-}
-
+// NOTE: This is likely to be created by the apps directly using `createStore` from redux,
+//       piecing together the reducers and middlewares they want
 export default function configureStore(initialState: AppState) {
   return createStore(
     rootReducer,
